requirements.txt
# Fistly install all via pip install --editable .
# Secondly install torch torchvision dependently
torch>=1.6.0, <=1.9.0
torchvision>=0.7.0, <=0.10.0
numpy>=1.16.6, <=1.21.4
tqdm>=4.43.0,<4.50.0
torchtext==0.5.0
GitPython==3.1.0
requests==2.23.0
fasttext==0.9.1
nltk==3.4.5
editdistance==0.5.3
transformers>=3.4.0, <=4.10.1
sklearn==0.0
omegaconf>=2.0.6, <=2.1
lmdb==0.98
termcolor==1.1.0
iopath==0.1.8
datasets==1.2.1
matplotlib==3.3.4
pycocotools
ftfy==5.8
pytorch-lightning #@ git+https://github.com/PyTorchLightning/pytorch-lightning@c335a7891
torchaudio>=0.6.0, <=0.9.0
psutil
pillow==8.3.1
setuptools==59.5.0
